refactor: clean up crates after element migration
- Update flui-element with stub methods for API compatibility
- Update flui-tree iterator and trait implementations
- Update flui-view wrappers and view_object module
- Update flui_animation, flui_app, flui_rendering dependencies
- Update flui_core pipeline files
- Add provider_view_object.rs and dirty_tracker.rs
- Remove migration.md and unified-renderobject-system openspec (completed)
- Update Cargo.lock with dependency changes

+++ crates/flui-tree/src/iter/render.rs
@@ -799,7 +799,7 @@ mod tests {
         }
 
         fn insert(&mut self, parent: Option<ElementId>, is_render: bool) -> ElementId {
-            let id = ElementId::new(self.nodes.len() as u64 + 1);
+            let id = ElementId::new(self.nodes.len() + 1);
             self.nodes.push(Some(TestNode {
                 parent,
                 children: Vec::new(),
