+++ crates/flui-tree/src/traits/combined.rs
@@ -189,7 +189,7 @@ mod tests {
         }
 
         fn insert(&mut self, node: TestNode) -> ElementId {
-            let id = ElementId::new(self.nodes.len() as u64 + 1);
+            let id = ElementId::new(self.nodes.len() + 1);
             self.nodes.push(Some(node));
             id
         }
