+++ crates/flui-tree/src/traits/nav.rs
@@ -499,7 +499,7 @@ mod tests {
         }
 
         fn insert(&mut self, parent: Option<ElementId>, slot: Option<Slot>) -> ElementId {
-            let id = ElementId::new(self.nodes.len() as u64 + 1);
+            let id = ElementId::new(self.nodes.len() + 1);
             self.nodes.push(Some(TestNode {
                 parent,
                 children: Vec::new(),
